+++ Domain/Domain.csproj
@@ -7,4 +7,9 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Metar.Decoder" Version="1.0.8" />
+    <PackageReference Include="Taf.Decoder" Version="1.0.5" />
+  </ItemGroup>
+
 </Project>
