+++ Domain/Reports/TafReport.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Taf.Decoder.entity;
 
-namespace MetarTaf_Backend.Models
+namespace Domain.Reports
 {
     public class TafReport : Report
     {
