+++ MetarTaf/MetarTaf.csproj
@@ -9,12 +9,12 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Metar.Decoder" Version="1.0.5" />
+    <PackageReference Include="Metar.Decoder" Version="1.0.8" />
     <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.20.1" />
     <PackageReference Include="Moq" Version="4.20.72" />
     <PackageReference Include="MoqProtectedLike" Version="2.0.1" />
     <PackageReference Include="Quick.AspNetCore.Components.Web.Extensions" Version="6.0.2" />
-    <PackageReference Include="Taf.Decoder" Version="1.0.2" />
+    <PackageReference Include="Taf.Decoder" Version="1.0.5" />
     <PackageReference Include="xunit" Version="2.9.0" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.8.2">
       <PrivateAssets>all</PrivateAssets>
