+++ MetarTaf_Backend/Models/IAirport.cs
@@ -1,4 +1,5 @@
-﻿using Domain.ValueObjects;
+﻿using Domain.Reports;
+using Domain.ValueObjects;
 using System;
 using System.Collections.Generic;
 using System.Linq;
