+++ MetarTaf_Backend/Models/IInfoStation.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Domain.Reports;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
