+++ MetarTaf_Backend/Services/MetarService.cs
@@ -1,4 +1,5 @@
-﻿using Metar.Decoder.Entity;
+﻿using Domain.Reports;
+using Metar.Decoder.Entity;
 using MetarTaf_Backend.Factories;
 using MetarTaf_Backend.Models;
